re-write over-exposure detectionalgorithm. Now it's based on Dimg Exposure settings and pure mask computation. New algorithm work as expected. 3 other detection algorithms still broken. Investiguations are under progress

+++ core/libs/dimg/filters/imgqsort/imagequalityparser.h
@@ -68,7 +68,11 @@ private:
     short  blurDetector2()            const;
     double noiseDetector()            const;
     int    compressionDetector()      const;
-    int    exposureAmount()           const;
+
+    /** Detect over-exposure amount in image. A pure white mask is computed and
+     *  a count of pure color pixels is used to evaluate the over-exposition of shot.
+     */
+    double exposureAmount()           const;
 
 private:
